requirements.txt
langchain
langchain-community
langchain-openai==0.0.8
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langgraph
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pypdf==3.17.1
tiktoken==0.5.2
faiss-cpu==1.7.4
python-multipart==0.0.6
langchain-core


